+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### App Center Distribute
 
-* **[Feature]** Add a **Disable Automatic Check For Update** checkbox that needs to be set in order to turn off automatic check for update. 
+* **[Feature]** Add a **Automatic Check For Update** checkbox that can be unchecked to turn off automatic check for update. 
 * **[Feature]** Add a `CheckForUpdate` API to manually check for update.
 
 ## Release 3.0.0